Moved the API methods to a separate file. Note that I had to move the `connexion_app.add_api()` call so that it happened *after* db.session was defined, so I could import the session into the api.py file
